GunCon 3 + negative gun-cal values

+++ input.h
@@ -99,7 +99,7 @@ uint32_t get_amiga_code(uint16_t key);
 uint32_t get_archie_code(uint16_t key);
 
 int input_has_lightgun();
-void input_lightgun_save(int idx, uint16_t *cal);
+void input_lightgun_save(int idx, int32_t *cal);
 
 void input_switch(int grab);
 int input_state();
